google-notion_api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
notion-client>=2.2.0
google-auth-oauthlib>=1.0.0
google-auth>=2.23.0
google-api-python-client>=2.100.0

